+++ storybook-static/assets/preview-78430b28.js
@@ -1,2 +1,0 @@
-import{_ as e}from"./iframe-887a8e89.js";import"../sb-preview/runtime.js";var a={docs:{renderer:async()=>{let{DocsRenderer:r}=await e(()=>import("./DocsRenderer-EYKKDMVH-14b4f585.js"),["./DocsRenderer-EYKKDMVH-14b4f585.js","./chunk-PCJTTTQV-764ac2f5.js","./iframe-887a8e89.js","./index-e0ddb630.js","./_commonjsHelpers-87174ba5.js","./index-d475d2ea.js","./index-d37d4223.js","./index-724aca61.js","./index-356e4a49.js"],import.meta.url);return new r}}};export{a as parameters};
-//# sourceMappingURL=preview-78430b28.js.map