+++ storybook-static/assets/syntaxhighlighter-QTQ2UBB4-b01a6a92.js
@@ -1,2 +1,0 @@
-import{S as l,c as n,s as d}from"./chunk-PCJTTTQV-764ac2f5.js";import"./iframe-887a8e89.js";import"../sb-preview/runtime.js";import"./index-e0ddb630.js";import"./_commonjsHelpers-87174ba5.js";import"./index-d475d2ea.js";import"./index-d37d4223.js";import"./index-724aca61.js";import"./index-356e4a49.js";export{l as SyntaxHighlighter,n as createCopyToClipboardFunction,d as default};
-//# sourceMappingURL=syntaxhighlighter-QTQ2UBB4-b01a6a92.js.map